Move id creation out of actor
When following along with the demo, I noticed that my ids were being recreated each time the app started, and they were never persisted.
I realized that this is because, like with redux, you can't define your ids in your reducers. They have to be part of the payload in the action creator from the beginning.

This was the most straightforward fix, although it's not as tidy.